+++ assets/directoried_codebase/CROMU_00007/README.cgc.md
