+++ assets/directoried_codebase/CROMU_00007/config.json
