+++ assets/directoried_codebase/CROMU_00007/crash.input
