+++ assets/directoried_codebase/CROMU_00007/pov/POV_00000.xml
