+++ assets/directoried_codebase/CROMU_00007/sample.input
