+++ assets/directoried_codebase/CROMU_00007/src/service.c
