+++ src/autopatchshared/autopatchshared/__init__.py
@@ -1,3 +1,4 @@
 from .get_current_timestamp import get_current_timestamp  # noqa: F401
 from .init_logging import init_logging  # noqa: F401
 from .load_config_from_json import load_config_as_json  # noqa: F401
+from .make_compile import make_compile  # noqa: F401
