+++ src/fuzzing-service/tests/pytest.ini
@@ -1,0 +1,4 @@
+# pytest.ini
+[pytest]
+asyncio_mode = auto
+asyncio_default_fixture_loop_scope = function
